iOS 10 fix for reflection process on block signatures.
For this class A2DynamicDelegate.m, specifically at this line (192):
NSMethodSignature *protoSig = [NSMethodSignature signatureWithObjCTypes:methodDescription.types];
For the argument of type MFMailComposeResult on iOS 9 the type definition is: "i" (int).
Running on iOS 10 the same line returns for the same argument: "q" (long long).

Later we have this method:
+ (BOOL)isSignature:(NSMethodSignature *)signatureA compatibleWithSignature:(NSMethodSignature *)signatureB __attribute_((pure))_
That method compares the argument between the block signature and the method signature. As the block reported the same argument described above as "i", on iOS 9 that condition is met because the arguments are same type. However on iOS 10 "i" is not equal to "q" and the methods returns FALSE. This cause the code throw the exception we are discussing in the thread.

I resolved for time being, creating a simple method to compare all the type definitions representing somehow an INTEGER value (https://developer.apple.com/library/content/documentation/Cocoa/Conceptual/ObjCRuntimeGuide/Articles/ocrtTypeEncodings.html#//apple_ref/doc/uid/TP40008048-CH100). So from that method I return true if an argument is "i" or "q", meaning both are integers. As I think Objective C can handle that difference later, it is working for me.